+++ tests/unit/test_utils_extract_namespace_name.py
@@ -2,7 +2,7 @@ from bigxml.utils import extract_namespace_name
 
 
 def test_no_namespace():
-    assert extract_namespace_name("foo") == (None, "foo")
+    assert extract_namespace_name("foo") == ("", "foo")
 
 
 def test_namespace():